requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
cryptography==2.2.2
Django==2.0.7
django-admin-rangefilter==0.3.7
django-suit==0.2.26
idna==2.7
mysqlclient==1.3.13
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
